+++ crates/fj-core/src/layers/objects.rs
@@ -2,7 +2,7 @@
 
 use crate::{
     objects::{AboutToBeStored, AnyObject, Objects},
-    validate::Validation,
+    validation::Validation,
 };
 
 use super::{Command, Event, Layer};
