+++ crates/fj-core/src/validate/mod.rs
@@ -73,18 +73,13 @@ mod solid;
 mod surface;
 mod vertex;
 
-use crate::{
-    storage::ObjectId,
-    validation::{ValidationConfig, ValidationError},
-};
+use crate::validation::{ValidationConfig, ValidationError};
 
 pub use self::{
     cycle::CycleValidationError, edge::EdgeValidationError,
     face::FaceValidationError, shell::ShellValidationError,
     sketch::SketchValidationError, solid::SolidValidationError,
 };
-
-use std::{collections::HashMap, error::Error, thread};
 
 /// Assert that some object has a validation error which matches a specific
 /// pattern. This is preferred to matching on [`Validate::validate_and_return_first_error`], since usually we don't care about the order.
@@ -97,54 +92,6 @@ macro_rules! assert_contains_err {
             errors.iter().any(|e| matches!(e, $p))
         })
     };
-}
-
-/// Errors that occurred while validating the objects inserted into the stores
-#[derive(Default)]
-pub struct Validation {
-    /// All unhandled validation errors
-    pub errors: HashMap<ObjectId, ValidationError>,
-
-    /// Validation configuration for the validation service
-    pub config: ValidationConfig,
-}
-
-impl Validation {
-    /// Construct an instance of `Validation`, using the provided configuration
-    pub fn with_validation_config(config: ValidationConfig) -> Self {
-        let errors = HashMap::new();
-        Self { errors, config }
-    }
-}
-
-impl Drop for Validation {
-    fn drop(&mut self) {
-        let num_errors = self.errors.len();
-        if num_errors > 0 {
-            println!(
-                "Dropping `Validation` with {num_errors} unhandled validation \
-                errors:"
-            );
-
-            for err in self.errors.values() {
-                println!("{}", err);
-
-                // Once `Report` is stable, we can replace this:
-                // https://doc.rust-lang.org/std/error/struct.Report.html
-                let mut source = err.source();
-                while let Some(err) = source {
-                    println!("\nCaused by:\n\t{err}");
-                    source = err.source();
-                }
-
-                print!("\n\n");
-            }
-
-            if !thread::panicking() {
-                panic!();
-            }
-        }
-    }
 }
 
 /// Validate an object
